requirements.txt
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
bitarray==3.8.0
certifi==2026.1.4
charset-normalizer==3.4.4
ckzg==2.1.5
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
cytoolz==1.1.0
debugpy==1.8.20
decorator==5.2.1
duckdb==1.4.4
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
executing==2.2.1
fastjsonschema==2.21.2
fonttools==4.61.1
h11==0.16.0
h2==4.3.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
narwhals==2.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.4.2
packaging==26.0
pandas==3.0.1
parsimonious==0.10.0
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
plotly==6.6.0
pluggy==1.6.0
poly_eip712_structs==0.0.1
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py_builder_relayer_client==0.0.1
py_builder_signing_sdk==0.0.2
py_clob_client==0.34.6
py_order_utils==0.3.2
pyarrow==23.0.1
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rlp==4.1.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
websockets==16.0
